updated soiling_hsu
